Custom Fields in Admin Screens & Investor

+++ DeepBlue/Models/Entity/Validation/Fund.cs
@@ -14,7 +14,7 @@ namespace DeepBlue.Models.Entity {
 			#region Primitive Properties
 
 			[Required]
-			[Range((int)(int)ConfigUtil.EntityIDStartRange, int.MaxValue)]
+			[Range((int)ConfigUtil.EntityIDStartRange, int.MaxValue)]
 			public global::System.Int32 EntityID {
 				get;
 				set;
