+++ DeepBlue/Models/Entity/Validation/InvestorEntityType.cs
@@ -11,7 +11,7 @@ namespace DeepBlue.Models.Entity {
 		public class InvestorEntityTypeMD {
 			#region Primitive Properties
 			[Required]
-			[Range((int)(int)ConfigUtil.EntityIDStartRange, int.MaxValue)]
+			[Range((int)ConfigUtil.EntityIDStartRange, int.MaxValue)]
 			public global::System.Int32 EntityID {
 				get;
 				set;
