+++ DeepBlue/Models/Entity/Validation/InvestorFundDocument.cs
@@ -28,7 +28,7 @@ namespace DeepBlue.Models.Entity {
 			}
 
 			[Required]
-			[Range((int)(int)ConfigUtil.EntityIDStartRange, int.MaxValue)]
+			[Range((int)ConfigUtil.EntityIDStartRange, int.MaxValue)]
 			public global::System.Int32 EntityID {
 				get;
 				set;
